Rename simulation shared to scenario_config.
