+++ src/plotting/task_plot_incidences.py
@@ -6,11 +6,11 @@ from src.config import BLD
 from src.plotting.plotting import plot_incidences
 from src.plotting.plotting import PY_DEPENDENCIES
 from src.policies.policy_tools import filter_dictionary
-from src.simulation.shared import (
+from src.simulation.scenario_config import (
     create_path_to_weekly_outcome_of_scenario,
 )
-from src.simulation.shared import get_available_scenarios
-from src.simulation.shared import get_named_scenarios
+from src.simulation.scenario_config import get_available_scenarios
+from src.simulation.scenario_config import get_named_scenarios
 from src.simulation.task_run_simulation import FAST_FLAG
 
 NAMED_SCENARIOS = get_named_scenarios()
